rename ocrd.log -> ocrd.utils to contain reusable static code
