+++ ocrd/utils.py
